Use standard buildProject from govuk_jenkins
This means our apps are all consistent with how they run tests on jenkins
but also that they gain all the benefits of changes to that script.  The
main blocker to this for whitehall was the custom args set while
precompiling assets, but [we merged those into the
script](https://github.com/alphagov/govuk-puppet/pull/6821/) so now we're
free to migrate over.