Remove the '[Archived]' title prefix from HTML attachments
https://trello.com/c/qZgjF7EN/472-rename-withdrawn-and-archived-html-publications

Some users have manually added '[Archived]' to the title of HTML
attachments, but the system already handles this. The result is we're
rendering titles like '[Withdrawn] [Archived] Something'.

At time of writing, there are 37 HTML attachments exhibiting this
problem. This is more than those listed in the Trello ticket, but we
decided to fix all of them.