Ensure opened files are closed
It is best practice to ensure that an opened file is closed using the
block form of `File#open` to free up the file descriptor.

On the development VM under an OS X host I noticed that before this
commit files starting `.nfs` would be created on the filesystem when
running these tests. In a subsequent commit I want to change the
implementation to glob for all files starting with `.`. This change
makes the subsequent one easier to test.