Regenerate locales to strip quotes
Changes to underlying versions of rails-i18n mean that when regenerating
translations using `rake translation:regenerate` we now have quotes
removed. Adding as it's own commit as a precursor to adding a new
translation key.

There is some irony in the fact that this is the opposite of a PR from
three years ago https://github.com/alphagov/whitehall/pull/2083.