Disable Style/MethodMissing cop
The excluded files in the Robocop configuration file contain a
`method_missing` method without also defining
`respond_to_missing?`. It would be preferable to introduce the
`respond_to_missing?` in each of the excluded files, however due to
time constraints I was unable to do so in this commit.

The cop will apply if any further `method_missing` methods are
introduced however additional work will be needed to address the
excluded files.