Pass relative paths to MigrateAssetsToAssetManager::Worker
See: https://github.com/alphagov/asset-manager/issues/414

Before this commit we were creating
`MigrateAssetsToAssetManager::Worker` jobs with the full path to each
asset to be migrated. Due to the way capistrono works and
`Whitehall.clean_uploads_root` is defined these paths took the form

    /data/vhost/whitehall-admin/releases/20180116155427/clean-uploads/system/uploads/person/1/image.jpg

If a deployment happened while these jobs were on the queue, they
would start to fail as the release directory would change.

With this commit we pass the relative path to the
asset (e.g. `system/uploads/person/1/image.jpg`) to the worker and
join it to the current `Whitehall.clean_uploads_root` when the job is
run.

I've changed the test to refute that `AssetFilePaths#file_paths` does
not return directories to use the absolute path. The way it was
written before it could have given a false positive.