Update Plek#public_asset_host in test
I found it confusing that the scheme of `public_asset_host` was `http`
but that we were asserting that the scheme of the `attachment_url` was
`https`.

It turns out that the scheme in `public_asset_host` is being ignored
because the `#attachment_url` method uses `Whitehall.url_maker` to
generate the URL. `Whitehall.url_maker` is an instance of
`Whitehall::UrlMaker` that's been instantiated using the protocol from
`Whitehall.public_protocol`.

I'm about to the change the implementation of `#attachment_url` and
changing this stubbed value now will mean that the test can stay
invariant when I change the behaviour of the method.