Publish localised routes for some pages
The announcements, publications, statistics, world-location-news and
ministers pages under /government can have .locale appended to the
path, where locale is a locale (e.g. es).

This commit changes 3 of these content items (statistics,
announcements and ministers) to have explicit routes for the localised
versions of the pages, rather than relying on the /government prefix
route.

This should help with removing the /government prefix route, something
I'm interested in doing, as I'd like to see it replaced with a taxon
rendered as a navigation page.