Remove 'placeholder' image alt texts
If no image has been selected for news articles, a placeholder image
gets added instead. And when that placeholder image is used,
its alt text reads 'placeholder'.
That is bad for screen reader users as that content doesn't make sense.
It's much better to leave it empty so screen readers can ignore it.