Add methods to AssetRemover to remove more asset directories
We are migrating ConsultationResponseFormUploader[1] and
ImageUploader[2] assets to Asset Manager which means they can be
removed from NFS. This commit adds methods to the AssetRemover class
to explicitly remove these directories. I've decided to add methods
rather than parameterise a public method because this is a destructive
action and being explicit about it feels safer.

[1] https://github.com/alphagov/asset-manager/issues/297
[2] https://github.com/alphagov/asset-manager/issues/405