Allow the disabling secure cookies in Rails production environment
The publishing end to end tests run Whitehall in Rails production
mode so we can mimic how the app will behave in production as closely
as possible.

We haven't set up HTTPS however which means that the secure cookie
marker is preventing the browser from storing the cookie, and
therefore sessions aren't working properly.

This ENV config allows us to run the app in Rails production mode
from end to end tests.