Implement AssetManagerStorage::File#content_type
The `ImageUploader#bitmap?` method expects the file returned by the
storage engine to implement a `content_type` method. This commit adds
one to the AssetManagerStorage::File class so that we can swap the
AssetManagerStorage engine into the ImageUploader.

When using the AssetManagerStorage we cannot directly access the file
on disk. When it is sent to Asset Manager it is uploaded to the
cloud and deleted from the file system. This means that we cannot
examine the file's content to determine the content type.

Instead we do what CarrierWave::SanitizedFile does[1] and use the
`mime-types` gem's `type_for` method to determine the content type
from the file extension.

[1] https://github.com/carrierwaveuploader/carrierwave/blob/v1.2.2/lib/carrierwave/sanitized_file.rb#L266