Fix 'updating should allow removal of images' test
The test setup using factories to create images was different to the
other examples in this shared group, and caused the ImageValidator to
fail as the uploaded file wasn't available to carrierwave. This didn't
happen when the app was run in development, only in the test, and I
didn't investigate further as using a consistent test setup fixed the
issue.

I also had to stub `whitehall_asset` to return an `id` so that the
AssetManagerDeleteAssetWorker would be able to "delete" the asset.