Fix 'updating an edition with an existing image allows image attributes to be changed'
This test was failing because when the caption or alt-text of the
image was changed the validator attempted to re-validate the attached
file. This file no longer exists as we now don't store assets on the
file system.

The slightly hacky fix for this is to return early from ImageValidator
if the file itself does not exist. This doesn't feel too bad as we're
already returning early if the ImageData object is not present and the
file will exist on initial upload (in the carrierwave tmp directory
before it is moved to asset manager).