Return basename from AssetManagerStorage::File#filename
The
`Admin::OrganisationsControllerTest#test_PUT_:update_can_set_default_news_image`
test was comparing the filename of the asset to an expected
filename. We were returning the full legacy_url_path which didn't
seem to match the intent of a method called `filename`. Instead we can
extract the filename part using File.basename().