Use asset manager storage for ImageUploader
Previously we were storing images uploaded using ImageUploader on both
the file system and in Asset Manager. We have now migrated all
existing ImageUploader assets to Asset Manager and have changed the
nginx configuration to serve requests for these assets from asset
manager. This means we no longer need to store new assets on the file
system.

This commit changes the storage engine to `asset_manager`. I've
changed the tests that verify that thumbnails of the correct size have
been created to use the file stored in `asset_manager_tmp_dir` before
it is sent to Asset Manager.

I've also removed tests that become redundant with this change. Virus
checking of images now happens in asset manager, so these integration
tests can be removed. We also no longer need to test in
AssetManagerIntegrationTest that assets are stored on the file system
- we have tests that cover the storage in Asset Manager.