Report files remaining after deletion
We were using `Dir.glob` to count the files deleted by these rake
tasks. It has two problems:

1. It counts directories as files
2. It takes around 40 minutes to count the files in large
directories (such as `image_data/files`).

We've verified that these tasks delete the files on integration so it
seems sufficient to just post a count after the files have been
deleted to check that they are gone.