Avoid prepending asset host to inline image URLs
This fixes the broken URLs we were generating when embedding inline
images into markdown.

We've recently updated the `ImageUploader` to use the
`AssetManagerStorage` engine. The `AssetManagerStorage::File#url` method
already includes the asset host so we don't need to prepend it again.

The problem was highlighted by this end-to-end test[1]

Note the addition of `require 'test_helper'` to
govspeak_content_worker_test.rb. This was so that I could run this test
in isolation using `bundle exec rails test`.

[1]: https://github.com/alphagov/publishing-e2e-tests/blob/master/spec/whitehall/uploading_attachment_spec.rb