Use correct image URL in CaseStudyPresenter
Since a024c0f8 we have been using the AssetManagerStorage engine to
store and serve images uploaded using the ImageUploader. This engine
returns `url` with the host already prepended to the path. This means
we shouldn't prepend the host in the CaseStudyPresenter. I've changed
the tests to match this implementation.