Use correct image URL in TakePartPresenter
Since a024c0f8 we have been using the AssetManagerStorage engine to
store and serve images uploaded using the ImageUploader. This engine
returns `url` with the host already prepended to the path. This means
we shouldn't prepend the host in the TakePartPresenter. Before this
commit the tests didn't catch this problem - I've added one so that we
prevent this issue again.