Mark Attachments as draft when saving them to Asset Manager
The ability to mark assets as draft was added to Asset Manager in
https://github.com/alphagov/asset-manager/pull/435.

I'm not particularly keen on this implementation but it gives me
something to test while I consider alternatives.

We're not going to be ready to serve these attachment assets from Asset
Manager for a while so marking them as draft seems like the safest
option in the meantime.