Pass the asset's draft status to the create worker
I've added the `assets_protected?` method to indicate whether the assets
should be protected when they're saved to Asset Manager. This is set to
false by default in the `WhitehallUploader` and overridden to be true in
the `AttachmentsUploader`. This should ensure that attachments are
marked as draft in Asset Manager, and that everything else continues to
be marked as non-draft (i.e. publicly accessible).