Stop using `params[:edition]` directly, prefer `edition_params`
Direction of travel for rails is to avoid using the params hash directly
and to only use params that have gone through the `permit`/`require` API
to make sure we don't touch things we shouldn't.

Note that we needed to add `attachment_action` to the list of permitted
params for consultations even though this is deleted and used to mutate
the params array and not ultimately sent to the edition.