Avoid calling `each_key` on editions_params
We can use `keys.each` to achieve the same thing and no longer rely on a
method on params that may not exist in rails 5.1 when params stop being a
hash.

Note that we have to disable the rubocop Performance/HashEachMethods cop
for this line because our intention is to stop using the each_keys method
it wants us to use.