Avoid calling `sort_by` on params in feature lists controller
This method may no longer exist on params objects in rails 5.1 when params
stop being a Hash.  To avoid calling it we need to make sure we've
permitted the relevant params, and then we can turn the object into a
hash and call sort_by on that directly.  To permit the relevant params we
check if the param name is a simple integer and reject it if not, then we
can safely permit the entire resulting params object.  For example given
params like:

    { ordering: { hats: 1, 1324: 0, 8245: 2, 234: 3 } }

we would end up with a permitted params object as follows:

    { 1324: 0, 8245: 2, 234: 3 }

We scope to just the `ordering` key, and we remove the keys in it that
aren't numbers.