Add asset_manager:migrate_attachments Rake task
This works slightly differently to the `asset_manager:migrate_assets`
Rake task in that it instantiates a `MigrateAssetsToAssetManager` object
for each attachment data directory instead of instantiating
a single `MigrateAssetsToAssetManager` for the parent attachment data
directory.

@chrislo reported that it took about 45 minutes just to queue up the
jobs to migrate Whitehall's images[1]. We think the vast majority of
this time was spent in `Dir.glob` as it had to build an array of about
500,000 files (about 50GB). The hope is that building an array of the
attachment_data directories (of which there are also about 500,000) and
queueing up jobs for each of these will give this task a better chance
of completing in a reasonable time.

[1]: https://github.com/alphagov/asset-manager/issues/404