PublicUploadsController only needs to redirect
There are no assets under Whitehall.clean_uploads_root except

1. Those in `uploaded/hmrc`
2. Attachments in `system/uploads/attachment_data`

Requests for 1. are handled by the HmrcAssetsController and requests
for 2. are handled by AttachmentsController.

This means the PublicUploadsController only has to redirect incoming requests
to the asset_host, so that they can be served by asset manager.

I've changed two tests in UploadAccessTest to assert that this
redirection takes place (it acts as an integration test of
routes.rb). The test in UploadAccessTest for the placeholder image is
redundant, Asset Manager now takes care of that.

Note that with the catch-all route for `/government/uploads/`[1] that
is currently in place it will be possible to craft a URL for an asset
that doesn't exist and cause the client to redirect infinitely. When
the final assets are migrated to asset manager, we can put a more
specific route in the nginx config to prevent this.

[1] https://github.com/alphagov/govuk-puppet/blob/7e6954cda9f7976d552faa12b0019eb054d4772b/hieradata/common.yaml#L1643