MigrateAssetsToAssetManager#migrate_attachments takes batch args
Allow MigrateAssetsToAssetManager#migrate_attachments to take range of
directory numbers to include in the migration.

This assumed, which is the case in practice on production, that the
directories under `attachment_data/file` are numbered, e.g.

    system/uploads/attachment_data/file/100
    system/uploads/attachment_data/file/200

We take the last part of the path, convert it to an integer and check
that it falls within the range.

There are no checks that the paths follow this convention, but I have
checked the list of current ones on production and they all do. Since
we are running the task interactively, I think it's OK to not be too
rigorous.