Make AM version of path available from AttachmentUploader
I want to add some functionality to update the draft status of an
attachment asset in Asset Manager. This will need to operate outside the
realms of CarrierWave and so I need to make this method available. We
should be able to remove it again once we switch the AttachmentUploader
to use the Whitehall::AssetManagerStorage class.