Update draft status in Asset Manager for attachment thumbnails
We now also enqueue a job for each of an attachment's thumbnails to
update the draft status of their associated assets in Asset Manager.

AttachmentData conditionally defines a thumbnail CarrierWave version for
PDF file types using the `if` option and the `#pdf?` method. However, it
appears that the thumbnail version is included in the Hash of versions
on the uploader whether or not `#pdf?` returns `true`; it's just that if
it's not a PDF, then the thumbnail version is not generated and is
therefore never `present?`.

Iterating through all versions and checking whether they are present
feels more robust than just optionally dealing with a single version
depending on the return value from the `#pdf?` method, because if
another version is added the code should continue to work.