Only update draft status in Asset Manager if has changed
Since we already have the `draft` status being returned in the response
from the call to `whitehall_asset` on the Asset Manager API, this
optimisation doesn't require the code to be much more complex and
although it might be slightly premature, I think it's worthwhile.

Note again that we're expecting the attribute keys to be strings,
because even if they are passed as symbols into
`AssetManagerUpdateAssetWorker.perform_async`, Sidekiq seems to convert
them into strings when they arrive at
`AssetManagerUpdateAssetWorker#perform`.