Fix consultation redirect params
This commit fixes the current implementation where the paramenters of
the url were being separated by a "." (dot) instead of a "?" (question
mark). It also replaced the ascii code "%255B" & "%255D" to the
appropriate ones which are ""%5B" & "%5D" (both mean "[" & "]") which
are used to represent an array of the attribute in cause.

This was brought to our attention in the Zendesk Ticket
https://govuk.zendesk.com/agent/tickets/2576393