Introduce AssetManagerUpdateAssetWorker class
I want to be able to update the draft status of attachment assets in
Asset Manager. This new worker class will allow me to do that.

It follows a similar pattern to AssetManagerCreateWhitehallAssetWorker
and AssetManagerDeleteAssetWorker. In particular, I've copied the code
for extracting the asset ID from its URL from the latter. Ideally this
duplciation would be extracted into
GdsApi::AssetManager#whitehall_asset, but that's for another day.

One other difference is that I decided to accept a Hash of attributes in
the `#perform` method in order to justify the relatively generic name of
the worker, even though for the moment I only need to pass the `draft`
attribute. Ideally we'd change the signature of
AssetManagerCreateWhitehallAssetWorker#perform from:

    perform(file_path, legacy_url_path, draft = false)

To:

    perform(legacy_url_path, attributes = {})

And pass the `file_path` & `draft` attributes in via the `attributes`
parameter, but I'm going to leave that for now, since changing it would
be complicated by need to cope with jobs that might already be in the
queue.