Prepare to update draft status in Asset Manager for attachments
This introduces the ServiceListeners::AttachmentDraftStatusUpdater class
which is intended to be wired up in
`config/initializers/edition_services.rb` to receive events whenever an
edition changes state. However, I haven't actually wired it up in this
commit and so the new class will have no effect. I plan to wire it up in
a subsequent commit.

The AttachmentDraftStatusUpdater iterates through all the attachments on
an "attachable" (an instance of a class which includes the Attachable
module) and uses an instance of the existing AttachmentVisibility class
to determine whether the attachment should be visible to the public.

For every attachment it queues up an AssetManagerUpdateAssetWorker job
to update the draft status of its corresponding asset in Asset Manager.

The AttachmentDraftStatusUpdater class doesn't yet support attachment
versions, i.e. PDF thumbnails. I plan to add support for this in a later
commit.

As it stands the AssetManagerUpdateAssetWorker will always update the
draft status whether it needs to or not. I plan to optimize this in a
later commit.

Note that we're using symbols as attribute keys in the call to
AssetManagerUpdateAssetWorker.perform_async, because it seems more
idiomatic Rails. However, Sidekiq seems to convert the keys into Strings
when the job is executed and AssetManagerUpdateAssetWorker#perform is
called. This is reflected in the unit test for
AssetManagerUpdateAssetWorker. I plan to introduce an integration test
in a subsequent commit which will ensure that this works as intended.