Update draft status in Asset Manager for attachments
This actually wires up the
ServiceListeners::AttachmentDraftStatusUpdater class to receive events
whenever an edition changes state. We're actually only interested in
changes of state which result in the visiblity of attachments changing
from not publicly visible to publicly visible or vice versa. However,
it's not trivial to work out exactly which events those are and so for
the moment we just subscribe to *all* events.

Wiring up the updater will mean that Asset Manager assets associated
with edition attachments will have their draft status kept up-to-date.
However, it's important to note that there are currently two other
non-edition classes (Response & PolicyGroup) which include the
Attachable module and therefore support attachments. These classes do
not themselves have state like the editions and we'll need to hook the
updater into the appropriate events in order to keep the draft status
for the assets corresponding to these attachments up-to-date. I plan to
address this in subsequent commits.

I've introduced an integration test which invokes the instances of
Edition::ForcePublisher and Edition::Unpublisher obtained from
Whitehall.edition_services in the same way as the actions on
Admin::EditionWorkflowController. I did consider introducing a Cucumber
feature to test this functionality via the UI, but since the only change
in behaviour is for the app to make calls to the Asset Manager API
rather than anything visible, it didn't seem worthwhile.

Note that the integration test is not intended to be exhaustive, but is
intended to capture a few of the more common scenarios in order to check
everything is wired up correctly.

Note that it's possible that subscribing to all events might cause an
undue number of AssetManagerUpdateAssetWorker jobs to be queued, but I'm
not convinced that that will be the case, and so I'm happy to avoid
premature optimisation until we discover that it is actually a problem.