Update draft status in Asset Manager for PolicyGroup attachments
This sets up an instance of
ServiceListeners::AttachmentDraftStatusUpdater as a subscriber to the
Whitehall.policy_group_notifier. This means that whenever a
policy group is created or updated, we iterate through all its
file attachments and update the draft status of their corresponding
assets in Asset Manager.

At the moment it's only actually possible to add attachments to a
persisted instance of PolicyGroup through the UI [1] and so I've only
added that scenario to the integration test. Having said that the code
should work OK if the UI was changed to allow the addition of
attachments at PolicyGroup create time.

[1]:
https://github.com/alphagov/whitehall/blob/88738c6be6024906ae52bdf5c25e299fe20c4ae5/app/views/admin/policy_groups/_form.html.erb#L12-L20