Update draft status of assets for consultation feedback attachments
It is possible to add a public feedback to a draft consultation and add
file attachments to that feedback. At this point the attachments should
not be visible to the public and so their corresponding assets in Asset
Manager should be marked as draft.

This was already working, because the assets for all new attachments are
created in the draft state [1]. However, if the parent consultation is
then published, the assets corresponding to the attachments on the
feedback were not being marked as published in Asset Manager.

This commit adds an extra call to
ServiceListeners::AttachmentDraftStatusUpdater#update! from the existing
edition_services event handler, but passing in the consultation feedback
instead of the ediiton.

I've added the scenario described above to the existing integration
test, but it's worth noting that there is a corresponding scenario where
a published consultation with a feedback with attachments is
unpublished.  However, I'm confident that the code will handle this
scenario and it didn't seem worthwhile adding another it to the
integration test.

[1]:
https://github.com/alphagov/whitehall/blob/88738c6be6024906ae52bdf5c25e299fe20c4ae5/app/uploaders/attachment_uploader.rb#L15-L17