Allow exclusion of england
Currently, when adding a publication you can indicate it to not be relevant to
Scotland, NI or Wales, it is always however relevant to England. For some Stats
publications, the document is actually only relevant to Scotland, This PR makes
it possible to make a document not relevant to England.