Set access_limited in Asset Manager when creating Whitehall assets
We want to be able to optionally restrict access to Whitehall assets in
Asset Manager to a subset of users; in the same way that we can restrict
access to a document in Whitehall.

I've modified the `AssetManagerStorage#store!` method to add the class
and ID (of the model that mounts the CarrierWave uploader) to the
`AssetManagerCreateWhitehallAssetWorker` job (e.g. `AttachmentData` or
`Organisation`). This allows me to retrieve the model within
`AssetManagerCreateWhitehallAssetWorker#perform` and ask it for any
access-limited restrictions it might have. I do this by checking whether
the model responds to the `attachable_is_access_limited?` method and
using the `PublishingApi::PayloadBuilder::AccessLimitation.for` method
to construct a set of users for the model's `access_limited_object`.

I've only added the `attachable_is_access_limited?` method to
`AttachmentData` at the moment. The `AttachmentData` class is used to
upload attachments to Whitehall and these are the only type of asset
that can be access limited. Instances of `AttachmentData` can be
associated with `Edition`s, `PolicyGroup`s and `Response`s. I've added
the `access_limited?` and `access_limited_object` methods to these
classes (note that `Edition#access_limited?` was already defined in the
`LimitedAccess` module) to provide a standard interface that
`AttachmentData` can use to query an object for its access-limited
information.

Note that this commit only deals with setting the access-limited data
when an asset is first uploaded to Asset Manager. I plan to make the
changes to keep this data up to date in subsequent commits.

Note that `AttachmentData#attachable` uses the most recently created
`Attachment` in order to lookup the attachable object. My assumption
here is that the most recently created attachment will belong to the
most recent edition (i.e. the one we care about when determining the
attachment visibility).

Note that `PublishingApi::PayloadBuilder::AccessLimitation.for` returns
a hash but I'm only interested in the array of users that are authorised
to access the attachment. I plan to extract this functionality in a
subsequent commit so that I can avoid overloading the purpose of the
`PublishingApi::PayloadBuilder::AccessLimitation` class.

Note that I've had to update a number of image_uploader and
attachment_uploader tests so that they create objects in the database.
This avoids the `AssetManagerCreateWhitehallAssetWorker#perform` from
raising an `ActiveRecord::RecordNotFound` exception during the tests. I
considered rescuing/ignoring this exception in the worker but I think
it'll be useful to be notified of any of these sort of exceptions when
the code is in production.

I initially thought I might be able to pass an array of authorised users
into the `AssetManagerCreateWhitehallAssetWorker#perform` method but
that proved problematic because we don't have all the required
information at the point we create that job in
`AssetManagerStorage#store!`. The `AttachmentData` object is saved,
which triggers the storage of the file, but this can/will happen before
the other objects have been saved which means we don't know which object
to ask for their access-limited status.