Only update draft status of attachments with underlying file in Rake task
I discovered that there are AttachmentData instances in integration with
no underlying file on the filesystem. This means that they have no
corresponding asset in Asset Manager and the asset lookup in
AssetManagerUpdateAssetWorker#perform would raise a GdsApi::HTTPNotFound
exception and the job would then be re-tried many times with no chance
of success. Since we haven't yet deleted the files from the Whitehall
part of the NFS mount, we can take advantage of this by not attempting
to update the asset draft status for attachments which have no
underlying file and thereby avoiding a bunch of failing jobs.