Update access_limited in Asset Manager when draft edition is updated
So that the access_limited state of assets attached to a draft document
are kept in sync with the same state on the parent document.

The new attachment_access_limited_integration_test.rb is a modified version
of attachment_draft_status_integration_test.rb, including its helper
methods. I imagine we'll remove this duplication in future.

I had to stub the linkables (using `publishing_api_has_linkables`) to
avoid some kind of JSON parsing error when one of the templates calls
`specialist_sector_options_for_select`.

I've updated the `AssetManagerUpdateAssetWorker` so that it only updates
both the `draft` and `access_limited` status when necessary. I imagine
we'll generalise this in future but this should do for now.