Simplify AttachmentDraftStatusIntegrationTest
There's no point in having all these tests check that the attachment
thumbnail asset has its draft status updated when the
ServiceListeners::AttachmentDraftStatusUpdaterTest already checks this.
No thumbnail is generated for non-PDF attachments and so we can simplify
the test accordingly. I think this makes it easier to follow what's
going on in the tests.