Accept redirect_url attribute in AssetManagerUpdateAssetWorker#perform
I want to be able to set the redirect_url for an asset in Asset Manager
in order to support the behaviour in AttachmentsController#fail which
redirects requests for attachments associated with an unpublished
edition to the parent document which should offer an explanation as to
why the document has been unpublished.

Although I only needed to add support to
AssetManagerUpdateAssetWorker#perform for the redirect_url attribute, I
decided it was just as easy to add support for arbitrary attributes.