Prepare to update redirect URL in Asset Manager for attachments
This introduces the ServiceListeners::AttachmentRedirectUrlUpdater class
which is intended to be wired up in
`config/initializers/edition_services.rb` to receive events whenever an
edition changes state. However, I haven't actually wired it up in this
commit and so the new class will have no effect. I plan to wire it up in
a subsequent commit.

The AttachmentRedirectUrlUpdater iterates through all the attachments on
an "attachable" (an instance of a class which includes the Attachable
module) and uses an instance of the existing AttachmentVisibility class
to determine whether the attachment should have a redirect URL.

For every attachment it queues up an AssetManagerUpdateAssetWorker job
to update the redirect URL of its corresponding asset in Asset Manager.

The AttachmentRedirectUrlUpdater class intentionally duplicates a lot of
code from AttachmentDraftStatusUpdater, e.g. it will also update the
redirect URL for the asset corresponding to a PDF attachment thumbnail.
I suspect that we will need to introduce other similar service listener
classes to support other asset-related behaviour and I think it's more
sensible to wait until then before deciding how best to remove the
duplication that this commit introduces.

Note that we're using symbols as attribute keys in the call to
AssetManagerUpdateAssetWorker.perform_async, because it seems more
idiomatic Rails. However, Sidekiq seems to convert the keys into Strings
when the job is executed and AssetManagerUpdateAssetWorker#perform is
called. This is reflected in the unit test for
AssetManagerUpdateAssetWorker. I plan to introduce an integration test
in a subsequent commit which will ensure that this works as intended.