Assets related to an unpublished document should not be marked as draft
In Asset Manager the draft status is used to decide whether to redirect
requests via the draft-assets host which requires authentication. The
fact that AttachmentVisibility#visible? returns `false` for attachments
associated with an unpublished document is a bit of an aberration,
because the attachment URL should in some sense be visible, i.e. it
should be accessible without authentication, it's just that requests to
it will be redirected to the URL of the parent document which will have
a public page explaining why the document has been unpublished.

This commit changes ServiceListeners::AttachmentDraftStatusUpdater so
that it does not mark assets for attachments associated with an
unpublished edition as draft.